Dashboard done
The debug box has been removed and it is now connecting to the robot. All sensors are added to network tables and the GUI is now pretty robust. The timer is connected to NT under match_time, the side view robot SVG can be controlled through '/SmartDashboard/hasCube' and '/SmartDashboard/elevatorPositionPercentage' with values from 0-100 (with 0 as lowest).